+++ RELEASE_NOTES.md
@@ -2,8 +2,8 @@
 
 [![Coverage Status](https://coveralls.io/repos/github/skuzzle/gh-prom-exporter/badge.svg?branch=master)](https://coveralls.io/github/skuzzle/gh-prom-exporter?branch=master) [![Twitter Follow](https://img.shields.io/twitter/follow/skuzzleOSS.svg?style=social)](https://twitter.com/skuzzleOSS)
 
-* Display application version
+* Support open metrics format (`Accept: application/openmetrics-text`)
 
 ```
-docker pull ghcr.io/skuzzle/gh-prom-exporter/gh-prom-exporter:0.0.6
+docker pull ghcr.io/skuzzle/gh-prom-exporter/gh-prom-exporter:0.0.7
 ```